pyproject.toml
[tool.poetry]
name = "djangochannels"
version = "0.1.0"
description = ""
authors = ["hbyyy <qjaduddl94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
django-extensions = "^2.2.9"
channels = "^2.4.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
